python/tools/mesh/requirements.txt
meshio
tetgen
libigl
polyscope
numpy
scipy
fast-simplification
gpytoolbox
pymeshfix